✨ Feat : 확인 창 구현
